fix: avoid tmux title conflict
If a tmux session already exists that begins with the name of the new
session, tmux will assume you want to "tab-complete" and open the
already present session.

For example, if a session `long-name` exist, and you want to create or
switch to `long`, so you first check if `tmux switch-client -t long`,
then tmux will open `long-name` instead of returning a failing exit
code.